+++ src/cri/src/lib.rs
@@ -9,6 +9,7 @@
 pub mod config_mapper;
 pub mod container;
 pub mod error;
+pub mod exec_request_id;
 pub mod image_service;
 pub mod persistent_store;
 pub mod runtime_service;